pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "datasets>=4.5.0",
    "detoxify>=0.5.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "google-cloud-storage>=2.19.0",
    "httpx>=0.28.1",
    "langchain>=1.2.4",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=0.4.0",
    "langchainhub>=0.1.21",
    "langsmith>=0.6.3",
    "langsmith-cli>=0.2.9",
    "matplotlib>=3.10.8",
    "pip>=26.0.1",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "python-multipart>=0.0.22",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "streamlit>=1.53.0",
    "streamlit-authenticator>=0.4.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E402",  # Module level import not at top — intentional in scripts & test helpers
    "E501",  # Line too long — pre-existing style in codebase
    "W293",  # Blank line contains whitespace — handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]          # Re-exports are intentional in __init__ files
"tests/giskard/*.py" = ["F401"]      # Optional giskard imports
"tests/promptfoo/*.py" = ["F401"]
"tests/integration/test_observability.py" = ["F401"]          # Optional observability imports
"src/content_analyzer/ner_detector.py" = ["F401"]             # Optional spacy type import
"src/content_analyzer/pii_detector_presidio.py" = ["F401"]    # Optional presidio import
"src/evaluation/*" = ["E722", "F841"]                          # Evaluation scripts — bare except and unused vars acceptable


[tool.coverage.run]
omit = [
    "src/evaluation/*",                        # Evaluation scripts — not unit-testable
    "src/content_analyzer/QUICK_START.py",      # Demo/quick-start script
    "src/content_analyzer/demo.py",             # Demo script
    "src/content_analyzer/utils.py",            # Standalone utility script
    "src/multi_embedding.py",                   # Standalone embedding script
    "src/observability/monitoring.py",          # Monitoring dashboard script
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",  # Exclude script entry points
    "pragma: no cover",
]

[tool.uv.sources]
torch = [{ index = 'pytorch-cpu' }]
torchvision = [{ index = 'pytorch-cpu' }]

[[tool.uv.index]]
name = 'pytorch-cpu'
url = 'https://download.pytorch.org/whl/cpu'
explicit = true

